Fix DCHECK in ComvertTimngForCompositor
When CovertTimingForCompositor was updated to use TimeDeltas instead of
doubles, a DCHECK(!infinity) was changed to a DCHECK that the scaled
time offset had an unsaturated value.  As large as the value is, it is
not infinite and legitimate values for animation arguments can trigger
a numeric overflow.  Instead, simply reject the animation on the
compositor.

Bug: 1054853
Change-Id: I941735b2ca098e111b39b2b99fbd32924b5d877b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080455
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Commit-Queue: Kevin Ellis <kevers@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746012}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cbf4dacfee89780299a389139b5aee8089e20e04