Add tracing event for long first input delay
A trace is uploaded via "slow reports" when a long first input delay is detected (for opted-in clients). This trace event will make it easier to identify the location of the long delay when analyzing the trace.

Bug: 1055890
Change-Id: I479406c3c78accbc4c06ccaeab14f699bf8ccdf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2078775
Reviewed-by: François Doray <fdoray@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Hailey Wang <haileywang@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#746526}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fbc6f8e75e04294b9edfd634ab320078895e8f02