DevTools: flush navigation events synchronously.
This allows associating navigation events with the input, the way it was
supposed to be at the first place.

By default, inspector events are coalesced and only flushed upon either
protocol command ack or blink task end. It is essential to not
compromise runtime performance. It is also essential to push notify in-time
for selected navigations and we already did that for many events, just
not the navigation ones. Fixing this.

Change-Id: I0b765bd634b3c03b64249150734e1368327b5b0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083562
Commit-Queue: Pavel Feldman <pfeldman@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746637}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f76cc49f1973842d50b5aa84d327298d20cd5d6