[CSP] Make SVGElement to handle "nonce" the same way as HTMLElement.
According to CSP, nonce are handled the same way for both HTMLElement
and SVGElement.

Both are setting the nonce when the Element is inserted, but only the
HTMLElement was supporting "modifying" a nonce.

It looks like a bug in Chrome found by annevk@:
https://github.com/web-platform-tests/wpt/pull/21853

This patch fixes the issue. It was meant to fix the WPT test:
- content-security-policy/nonce-hiding/nonces.html

But it turns out it is also fixing two more tests
- content-security-policy/nonce-hiding/svgscript-nonces-hidden.html
- content-security-policy/nonce-hiding/svgscript-nonces-hidden-meta-sub.html

Bug: 1053496
Change-Id: I872cae74817bff2f5f910dcd7864fc97426c49cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2075340
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746774}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c70f383897c785c804fa7e7df2bbc1d004a3b973