[devtools] Add inspector protocol test for new COEP error reporting
The error reporting was landed recently here:
https://chromium-review.googlesource.com/c/chromium/src/+/2070409

This CL adds a layout test that checks that the new errors are actually
reported via the Chrome DevTools Protocol (CDP).

Note that this test relies on the --cross-origin-isolation flag to
be turned on, which it is for layout tests, and we don't expect it
to be turned off. If the flag gets turned off, this test will fail
and needs to be disabled.

Bug: chromium:1051473
Change-Id: Iaccb843923ae1b2654ea11f3eec3f5b04bd1d551
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2087913
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#746779}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10c9e1cb4b65ef0dd172d94df7e938750f47c866