Reland "Fix null check preventing AXObjectCacheImpl from registering as popup document lifecycle observer"
The change was reverted because a couple of the updated tests were failing
on older mac platforms.  Apparently opening the date/week popops fires
different numbers of events on these platforms.  Since the point of
the tests is to observe events fired from focus changes in an opened
popup, I changed the tests to start logging events only after opening
the popup.  I've also modified them to provide some logging in the event
of a failure instead of just timing out opaquely.

Initial change description follows:

Fix the polarity on a couple of null checks that were preventing
AXObjectCacheImpl from registering as a lifecycle notification observer
for popup documents.  This failure to register was preventing AX focus
events from being fired when focus changed inside of a popup.

The week-picker-ax.html test was passing prematurely because of focus
events received when opening the popup, without validating that
additional events were fired when changing focus in the popup.  Update
that test to confirm that events are also fired for a focus change in
the popup itself.

There were a couple of analogous tests (date-picker-ax.html,
month-picker.html) that were deleted when we purged old date popup tests
in favor of new tests written for the new popup implementation
(crrev.com/c/1967608).  I don't see replacements for these in the new
tests, so this change also restores those tests, updates them for
the structure of the new form controls, and ensures that they are
checking that focus events are fired due to changes in the popups and
not just from events fired when the popups open.

TBR=aboxhall@chromium.org

Bug: 1022920, 1057036, 1057044
Change-Id: I8792ea074e89967c7ff68e21dfda8e0f6ba6f9a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080459
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#747012}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b4452765bcb6096aba5e4d66f3dda902aee08d0