Use more sophisticated faking mechanism for tests using WebMediaPlayer
Previously, the way time progression was simulated for
FakeWebMediaPlayer in tests was to increment the current time by a whole
second every time a call to |FakeWebMediaPlayer::CurrentTime| was made.
This is kind of strange, since it relies heavily on the implementation
of |HTMLMediaElement| to not call |CurrentTime| more than once per
update. Additionally, it led to scenarios where time in
FakeWebMediaPlayer would progress more quickly than test time (such as
in HTMLMediaElementWithMockSchedulerTest.PeriodicTimeupdateAfterSeek).

The new mechanism uses a recurring delayed task to increment time, and
also calls |Microtask::PerformCheckpoint| in order to update
|HTMLMediaElement::official_playback_position_|. Previously, tests would
fail unless they were specifically designed to avoid code paths that
rely on that value of |HTMLMediaElement::official_playback_position_|
(which ideally shouldn't be necessary). This CL resolves all of the
above issues, allowing some tests to be simplified (such as
HTMLMediaElementWithMockSchedulerTest.CueEnterExitEventLatency, which
is implemented in a pending CL: 2008079).

Change-Id: Ia0392a77a013b33a492f9a29402605ceb7f147ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2048415
Commit-Queue: Will Cassella <cassew@google.com>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#747075}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2dba6fd3454dac04455e2be6d80dbe300ff1d2de