Move highlight-on-hover for calendar picker to CSS
This change moves the logic for handling highlight-on-hover for cells
in date controls out of the JS and into :hover CSS rules.  I'd
originally planned for this work to wait until the old form controls
implementation was removed.  However, the highlight-managing code is
setting ARIA attributes in a way that doesn't make sense now that the
highlight positions no longer drive selection and are merely decorative.
So as part of improving the screen reader experience for these controls
it seems best to go ahead and do this work now.

The highlight-managing code is now all effectively dead code when
FormControlsRefresh is enabled that can be straightforwardly
chopped out when the time comes to remove the old implementation.

A follow-up change will ensure that the correct ARIA attributes are
set for selection changes.

Bug: 1046054, 1053798
Change-Id: Ie5081ef23b308a82bfc03887f96c912a82f411b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2083866
Commit-Queue: Dan Clark <daniec@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#747984}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 90adbbfff15527be854ca20f301d79dc11906332