system-web-apps: add support to enable origin trials
This CL adds support for system web apps to enable origin trials. It
uses WebAppTabHelper to enable origin trials on ReadyToCommitNavigation.

Works when launching SWA from shelf or when being navigated to in a tab.

Under the hood, it adds a new ForceEnableOriginTrials method to
NavigationHandle, which sets the origin trials for this navigation.
Internally, the trials are saved to commit_params.

When RenderFrame commits this navigation, it passes the origin trial
information to blink via WebNavigationParams. The DocumentLoader then
enables the provided trials.

Fixed: 1024655,1043842
Change-Id: I4030575c21454c0addb4953992149cdd4ee6ffe0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2038511
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Jason Chase <chasej@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Commit-Queue: Jiewei Qian  <qjw@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748120}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0406fc0b99990567551a33ba7634d95341ebe577