[cascade] Introduce surrogate properties
Non-alias-properties that share a storage location on ComputedStyle with
some other property pose a problem for StyleCascade, since we must
preserve the ordering between them. For example:

  div {
    width: 10px;
    inline-size: 20px;
  }

In the above, we do not know if inline-size refers to width or height
until the computed value of direction and writing-mode are known.
Therefore, inline-size exists in the cascade alongside width, and
the "winner" must be determined not during Analyze, but during Apply.

In the first StyleCascade design, StyleCascadeSlots were used for this
purpose, which acted as a second level cascade. The first property to
be applied (width or inline-size) would effectively block the other
one from applying. This worked, but it was too slow: it has a negative
effect on certain microbenchmarks.

The problem is that we have to do extra work for regular properties
(like width), even when e.g. inline-size isn't used.

With the new StyleCascade design, it makes more sense to just skip
'width' if we're applying 'inline-size' with a higher priority.

Bug: 947004
Change-Id: Ibe3af34204dff226379e30dca17d93b337d0513f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2080421
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748237}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33e1ab7993e38c478373782fed67d8a3ba786465