Throw an exception when ExecutionContext is not available
Some AudioNodes require an access to the execution context (mainly
because of the task runner) and this might not work (null-deref) if
the associated frame/document of a BaseAudioContext is already
detached.

This change adds a check against ExecutionContext in the factory
methods of problematic cases.

Test: Ran submitted reprocases on ASAN and they do not crash anymore.
Bug: 1057735, 1058693, 1058648, 1058687, 1058668
Change-Id: I4c6b994087f8f9561edd4fe8eaa74961fa3bebbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095141
Commit-Queue: Hongchan Choi <hongchan@chromium.org>
Reviewed-by: Raymond Toy <rtoy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748478}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 264d876a4febe92e83e91976ec7f302cac28fccd