Implement composite order with animate preempting css
In this patch, we replace the current implementation of sorting effect
order by comparing effect sequence number with comparing the
HasLowerCompositeOrdering function.

The patch will align the implementation with the spec
(https://www.w3.org/TR/web-animations-1/#the-effect-stack). And as we
are using the same comparison function as the getAnimations, the
rendered result would be the same as getAnimations() result.

Bug: 1050224

Change-Id: I88f151fa78070817d2fa4fd520580ddf6d024e27
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2076160
Commit-Queue: Hao Sheng <haozhes@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748492}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93d5df8ba14f8c9d3abc19ff42af7299eaa8bd29