Add |NextIntervalPoint| method to |WTF::PODIntervalTree|
This method is useful for querying when the next "interval point"
(start or end) is from a given starting point. Which is quite handy
if you're using this class to represent some sort of timeline...

Speaking of which, this change was done to support
https://chromium-review.googlesource.com/c/chromium/src/+/2008079

It also cleans up the code a fair amount as well.

Change-Id: I2f9fc6fdce9ae8cb05264f73f4d392e79cc8e7a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2073058
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#748513}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f81408a35815598a14f8b8c9065f6e06bab17d1