Change the meaning of symbol_level=0 on Windows
symbol_level=1 on Windows was changed to include filename and line
number information which gives us better call stacks. However on Windows
7 the symbol lookup became much slower, leading to timeouts. This change
modifies symbol_level=0 so that it includes function names (but no
filenames or line numbers) so that it is what symbol_level=1 used to be.
The build machines can then selectively be changed to symbol_level=0 to
avoid the timeouts, while still giving higher quality call stacks when
symbol_level=1 is selected.

When auditing the .gn files to validate this files it is important to
search both for symbol_level == 0 and symbol_level != 0. This may still
miss some cases but I think the changes are correct.

The total elapsed time sum for linking all 330 PEFile targets in a debug
component build of the 'chrome' target at symbol_level=0, as reported by
post_build_ninja_summary.py, goes from 250 s to 325 s with this change.
The weighted time (an estimate of the impact on build time) goes from
~25 s to ~39 s with this change, so a 14 s increase in build time.
The total size of the PDBs generated in this scenario is about 2.1 GB.

Bug: 985255
Change-Id: I1ea25433a1cf22f906c493ee0df1e41c98ed923d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1986973
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Reid Kleckner <rnk@chromium.org>
Commit-Queue: Bruce Dawson <brucedawson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748527}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3210c62d6b7d1878890166fe5f3dadf281dc1922