FontAccess: FontCache-based enumeration of Mac system fonts
Implements the enumeration of system fonts on Mac.
This change calls the CoreText APIs to enumerate the list of fonts
installed on a system.

It has been implemented in FontCache so that it
is accessible in the renderer process due to concerns about expensive
I/O if implemented browser-side.

Caching code will come in a subsequent CL.

Bug: 1043306
Change-Id: Iae9ce6c08174d8b4ced80eda07c795444998c746
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2091597
Commit-Queue: Olivier Yiptong <oyiptong@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748560}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9fc1063c31d67c3137ce2771060d2a62f17b808c