Use Promise with RequestPointerLock calls
The Pointer Lock API used to use an even based system.  However, that
is an outdated system and the events do not give an reason in failures.
This change introduces a Promise to the call that adds reasons if the
request is rejected.

Bug: 1042289
Change-Id: I4c8cacac3db5d9dc8d2f3a09c663c2dfb5bdf89e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2069199
Commit-Queue: James Hollyer <jameshollyer@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748810}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d5c9de46626708f710bed9634e1c598d4ac506fa