Introduce scheduler.currentTaskSignal
This returns the TaskSignal that was passed to scheduler.postTask()
when the current task was scheduled in the general case. When outside
of a posted task or when a custom priority was provided to postTask(),
it creates an implicit TaskSignal encoding the relevant state for the
posted task.

Bug: 979017
Change-Id: I9f960fb76b2500dbc851971e2d9bd2c0f9c51567
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2079386
Commit-Queue: Nate Chapin <japhet@chromium.org>
Auto-Submit: Nate Chapin <japhet@chromium.org>
Reviewed-by: Scott Haseley <shaseley@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748852}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fca88beb2d95fbea1f1bb59fd4629ae6933e7a38