bluetooth: getDevices() implementation
This change implements getDevices() which returns a list of
WebBluetoothDevice objects that the current site has permission to
access. If the kWebBluetoothNewPermissionsBackend flag is enabled, the
list of devices will contain all of the permitted devices. If the flag
is not enabled, then the list of devices will contain the permitted
devices that are currently connected to the system.

Design doc:
https://docs.google.com/document/d/1h3uAVXJARHrNWaNACUPiQhLt7XI-fFFQoARSs1WgMDM/edit#heading=h.5ugemo7p04z9

Bug: 577953
Change-Id: I9785f24ee46ac634b6a96d6146f54da37d132a4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2044660
Commit-Queue: Ovidio de Jesús Ruiz-Henríquez <odejesush@chromium.org>
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Reviewed-by: Vincent Scheib <scheib@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#748917}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 278801c2cebadcca72e9b8534114db067011944d