Delete duplicate operator<< for InsecureRequestPolicy.
crrev.com/c/2071164 created an unusued operator<< for this type
which conflicts with the autogenerated one in the out/ directory.

In most builds this is OK because the bad definition is `inline`, but
never used, so the compiler never emits into an object file. However
in a coverage build the symbol has to be emitted so that coverage can
be tracked, and then it collides with the other operator<< definition.

Note that it's illegal to have both inline and non-inline definitions
of the same function in C++, so this was always invalid code, even
though it took a behavior of the coverage system to expose it.

Bug: 1060343
Change-Id: Ifb5f5143fa5d651e17305ba1b364b22af644b93b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2097128
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Sajjad Mirza <sajjadm@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76e420b8cf853ed15771ed4bc2653b768a74d447