[RTCInsertableStreams] Use WTF::Vector instead of std::vector in RTCEncodedVideoFrame
Switch the additional_data_vector_ field to WTF::Vector for better
compliance with Blink practices.
While this implies a copy of the data, this vector is normally small
and we plan to change it to a structured format in the near future.

Bug: 1059912
Change-Id: I0fdeb5fefd4aac9c85712bf246866e5826b5f2c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2094311
Commit-Queue: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Marina Ciocea <marinaciocea@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749254}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 536a6d3945a4a030bd81028963498c0c32afe49c