Fixes Next/PreviousOnLine when an inline spans multiple lines
When trying to compute the next or previous AXObject on the same line, and when not using LayoutNG,
and when an AXLayoutObject is associated with an inline layout object which doesn't have any of its siblings
on the same line, we need to recursively look at the parent AXLayoutObjects.
However, before doing that we should ensure that the current AXLayoutObject is either the first or the last child
because any of our ancestors may span multiple lines.
Also, we should ensure that if "culled inline" is in effect, we will still walk the inline line boxes.

In addition, This patch fixes an issue with associating a list marker with
its list item via a next/previousOnLine relationship, when
the list item doesn't have any visible text, but only accessible text, e.g. an aria-label.

Example HTML snippet:
<div contenteditable="true">
  This is a
  <a href="example.com">
    link
    <span style="outline-style: solid">
      test with a
      <br>
      second line.
    </span>
  </a>
</div>

Observed issue:
All the contents of the contenteditable will appear to be on a single line.

Cause:
Chromium's renderer uses line boxes to represent each line.
There were two mistaken assumptions in AXLayoutObject:
1. Assuming that line boxes are always created.
Line boxes are not always created for every inline element; an optimization known
as "culled inline".
Only if the element has margin, decoration, padding, and outline, will
line boxes be created.
2. Assuming that if a node in the accessibility tree has no other nodes either before or after it
on a given line, then we should always look for such nodes in the siblings of its
parent object. However, this won't work if the parent node represents an inline element,
because that element can span multiple lines.
For example, a span on the first line will be marked as being on the same line as
a text node on the second line, if there is a second span that
starts on the first line but finishes on the second line, and which contains
the text node.

R=aleventhal@chromium.org, dmazzoni@chromium.org, kbabbitt@microsoft.com

Change-Id: I8815eeed629de0949fce598e9443163045ff7570
Bug: 1059612
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2090102
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 51371b2f2ed4d5937ef7430a0caae51aa395a47f