heap: Avoid data races when accessing bit fields for TraceAfterDispatch
This is a followup to CLs 2074881 and 2078533.
The flags used for TraceAfterDispatch are const and are never changed
after initialization. However, because these are bit-fields and each
of the flags is only a few bits wide, they share their memory with
other mutable fields. As a result, TSan generates warnings when reading
these flags.

Bug: 1059850
Change-Id: I4c1f216ccfc82e36d9066a6643bd271e3234eefd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2096450
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749417}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4bba1bd9f533e3f12b2e50bc94a784f7219c3a99