[NativeFS] Add origin trial integration for second OT.
Renames the existing trial to NativeFileSystem2, and make the same trial
name also trigger writable streams and cloneble handles.

Finally this also removes (the bindings for) the old write API, as this
API is not supposed to be part of the new origin trial. The implementation
of that method will be removed in future CLs.

Bug: 853326
Change-Id: I01bc2e091fed1c7036086783ad9ac713f0656be1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095437
Reviewed-by: Victor Costan <pwnall@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Commit-Queue: Marijn Kruisselbrink <mek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749464}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ad81e6202d120884fcc649002424ad543d5d5d1b