Use unsigned Web Bundle claimed URL for window.location and document.URL
Currently the claimed URL of unsigned Web Bundle file is used only for
relative path computation in the document. This is done by setting the
base URL of the document. (https://crrev.com/c/1802902)
  Note: |base_url_override_for_bundled_exchanges| was renamed to
  |base_url_override_for_web_bundle| by https://crrev.com/c/1923786. And
  this CL renames it to |web_bundle_claimed_url|.

But the file path of the Web Bundle file is still available using
window.location and document.URL. This is problematic from the privacy
point of view. This may leak the user name.
  Note: The same problem exists in opening local HTML files
  (https://crbug.com/990216).

To solve this problem, this CL changes window.location and document.URL
to use the claimed URL. This behavior is written in the explainer doc of
“Navigation to Unsigned Web Bundles”.
https://github.com/WICG/webpackage/blob/master/explainers/navigation-to-unsigned-bundles.md#loading-an-untrusted-bundle

Bug: 995177,1023929

Change-Id: I922c4b40504fc1ed87e78cf77c6877b59110dedc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2040500
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749646}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8878baa236d4dbb3dc2ebb565670dc7c301a4088